# ChatDiffieHelman.Api/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "description": "Аутентификация пользователя с email и паролем. Может потребовать дополнительную идентификацию (Fiat-Shamir или BMC).",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход или требуется дополнительная идентификация",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "nullable": true,
                      "description": "JWT токен (null если требуется дополнительная идентификация)"
                    },
                    "fiat_required": {
                      "type": "boolean",
                      "description": "Требуется ли идентификация Fiat-Shamir"
                    },
                    "fiat_session_id": {
                      "type": "string",
                      "nullable": true,
                      "description": "ID сессии Fiat-Shamir"
                    },
                    "bmc_required": {
                      "type": "boolean",
                      "description": "Требуется ли идентификация Brickell–McCurley"
                    },
                    "bmc_session_id": {
                      "type": "string",
                      "nullable": true,
                      "description": "ID сессии BMC"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Неверные учетные данные"
          }
        },
        "summary": "Вход пользователя",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/uniauth/start": {
      "post": {
        "description": "Получает Token 1 в UniAuth и возвращает redirectUrl на bridge страницу UniAuth.",
        "operationId": "AuthController_startUniAuth",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ссылка для перехода на UniAuth bridge",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "redirectUrl": {
                      "type": "string",
                      "example": "http://localhost:4200/oauth2/external-redirect/<token1>"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Начать вход через UniAuth",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/uniauth/callback": {
      "get": {
        "description": "Принимает token3, валидирует его через UniAuth и редиректит пользователя обратно во внешний frontend.",
        "operationId": "AuthController_uniAuthCallback",
        "parameters": [
          {
            "name": "token3",
            "required": false,
            "in": "query",
            "description": "Одноразовый Token 3 от UniAuth bridge",
            "schema": {}
          }
        ],
        "responses": {
          "302": {
            "description": "Редирект на login страницу web-клиента с результатом входа"
          },
          "502": {
            "description": "Fallback HTML-страница при ошибке redirect или сети"
          }
        },
        "summary": "Callback входа через UniAuth",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/fiat/start": {
      "post": {
        "description": "Первый шаг протокола идентификации Fiat-Shamir",
        "operationId": "AuthController_fiatStart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "sid": {
                    "type": "string",
                    "description": "ID сессии"
                  },
                  "t": {
                    "type": "string",
                    "description": "Обязательство (commitment) t = r² mod n"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Возвращает вызов (challenge)",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "c": {
                      "type": "number",
                      "description": "Случайный вызов (0 или 1)"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Начать идентификацию Fiat-Shamir",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/fiat/finish": {
      "post": {
        "description": "Второй шаг протокола идентификации Fiat-Shamir",
        "operationId": "AuthController_fiatFinish",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "sid": {
                    "type": "string",
                    "description": "ID сессии"
                  },
                  "r": {
                    "type": "string",
                    "description": "Ответ r = x·c + y mod n"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT токен при успешной идентификации",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "nullable": true,
                      "description": "JWT токен доступа"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Идентификация не пройдена"
          }
        },
        "summary": "Завершить идентификацию Fiat-Shamir",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/bmc/start": {
      "post": {
        "description": "Первый шаг протокола идентификации Brickell–McCurley",
        "operationId": "AuthController_bmcStart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "sid": {
                    "type": "string",
                    "description": "ID сессии"
                  },
                  "a": {
                    "type": "string",
                    "description": "Обязательство a = g^k mod n"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Возвращает случайный вызов",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "r": {
                      "type": "number",
                      "description": "Случайное число вызова"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Начать идентификацию Brickell–McCurley",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/bmc/finish": {
      "post": {
        "description": "Второй шаг протокола идентификации Brickell–McCurley",
        "operationId": "AuthController_bmcFinish",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "sid": {
                    "type": "string",
                    "description": "ID сессии"
                  },
                  "e": {
                    "type": "string",
                    "description": "Ответ e = k + r·x mod (n-1)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT токен при успешной идентификации",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "nullable": true,
                      "description": "JWT токен доступа"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Идентификация не пройдена"
          }
        },
        "summary": "Завершить идентификацию Brickell–McCurley",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/fiat/enable/{userId}": {
      "post": {
        "description": "Активация протокола Fiat-Shamir идентификации",
        "operationId": "AuthController_enableFiatForUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "v": {
                    "type": "string",
                    "description": "Публичный ключ v = s² mod n"
                  },
                  "n": {
                    "type": "string",
                    "description": "Модуль n (произведение двух простых чисел)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Сообщение об успехе",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Включить Fiat-Shamir для пользователя",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/fiat/disable/{userId}": {
      "post": {
        "description": "Деактивация протокола Fiat-Shamir идентификации",
        "operationId": "AuthController_disableFiatForUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Обновленные данные пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Отключить Fiat-Shamir для пользователя",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/bmc/enable/{userId}": {
      "post": {
        "description": "Активация протокола Brickell–McCurley идентификации",
        "operationId": "AuthController_enableBmcForUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "n": {
                    "type": "string",
                    "description": "Модуль n"
                  },
                  "g": {
                    "type": "string",
                    "description": "Генератор g"
                  },
                  "y": {
                    "type": "string",
                    "description": "Публичный ключ y = g^x mod n"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновленные данные пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Включить Brickell–McCurley для пользователя",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/bmc/disable/{userId}": {
      "post": {
        "description": "Деактивация протокола Brickell–McCurley идентификации",
        "operationId": "AuthController_disableBmcForUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Обновленные данные пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Отключить Brickell–McCurley для пользователя",
        "tags": [
          "auth"
        ]
      }
    },
    "/chats": {
      "get": {
        "description": "Возвращает список всех чатов текущего пользователя",
        "operationId": "ChatController_getMyChats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список чатов",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Получить мои чаты",
        "tags": [
          "chats"
        ]
      },
      "post": {
        "description": "Создает новый чат с указанным названием",
        "operationId": "ChatController_createChat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Название чата",
                    "example": "My Chat Room"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Созданный чат",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Создать новый чат",
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/{chatId}": {
      "get": {
        "description": "Возвращает информацию о конкретном чате",
        "operationId": "ChatController_getChat",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "UUID чата",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Данные чата",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Чат не найден"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Получить чат по ID",
        "tags": [
          "chats"
        ]
      }
    },
...
  },
  "info": {
    "title": "Chat Diffie-Hellman API",
    "description": "\n      API для чата с криптографической защитой.\n      \n      ## Возможности:\n      - 🔐 Аутентификация JWT\n      - 🔑 Протокол Fiat-Shamir идентификации\n      - 🛡️ Протокол Brickell–McCurley идентификации  \n      - 💬 Управление чатами и сообщениями\n      - 🔒 Обмен ключами Diffie-Hellman для шифрования\n      - 👥 Система приглашений\n      \n      ## Как начать:\n      1. Зарегистрируйтесь через `POST /users/registration`\n      2. Войдите через `POST /auth/login`\n      3. Используйте полученный токен для авторизации (кнопка \"Authorize\" вверху)\n      \n      ## JSON документация доступна по адресу:\n      - Swagger UI: `/api`\n      - JSON схема: `/api-json`\n      - YAML схема: `/api-yaml`\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "description": "Введите JWT токен, полученный после авторизации"
      }
    },
    "schemas": {
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "ID пользователя (UUID)",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "Имя пользователя"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "Email пользователя"
          },
          "fiat_enabled": {
            "type": "boolean",
            "example": true,
            "description": "Включена ли функция Fiat"
          },
          "bmc_enabled": {
            "type": "boolean",
            "example": false,
            "description": "Включен ли протокол Brickell–McCurley"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "fiat_enabled",
          "bmc_enabled"
        ]
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "Email пользователя"
          },
          "password": {
            "type": "string",
            "example": "strongPassword123",
            "description": "Пароль пользователя"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ChatDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "ID чата (UUID)",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Chat Room 1",
            "description": "Название чата"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
    }
  },
  "security": [
    {
      "access-token": []
    }
  ]
}